requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
